requirements.txt
gunicorn
django
django-allauth
libsass
django-compressor
django-sass-processor
oauth2client
gspread